v5

+++ app/Http/Controllers/Admin/Location/LocationController.php
@@ -17,33 +17,35 @@ use Yajra\DataTables\Facades\DataTables;
 
 class LocationController extends Controller
 {
-    public function index(){
-        $category_contents=CategoryContent::query()->where('type','location')->select('uuid','name')->get();
-        $users=User::query()->select('name','uuid')->get();
+    public function index()
+    {
+        $category_contents = CategoryContent::query()->where('type', 'location')->select('uuid', 'name')->get();
+        $users = User::query()->select('name', 'uuid')->get();
 
-        return view('admin.locations.index',compact('category_contents','users'));
+        return view('admin.locations.index', compact('category_contents', 'users'));
     }
+
     public function store(Request $request)
     {
         $rules = [
             'name' => 'required|string|max:36',
             'price' => 'required|int',
             'details' => 'required',
-            'user_uuid'=>'required|exists:users,uuid',
-            'category_contents_uuid' => 'required|exists:category_contents,uuid',
+            'user_uuid' => 'required|exists:users,uuid',
         ];
-
         $this->validate($request, $rules);
-        $location= Location::query()->create($request->only('name','user_uuid','price','details','category_contents_uuid'));
+        $location = Location::query()->create($request->only('name', 'user_uuid', 'price', 'details'));
+        $location->categories()->sync($request->category_contents_uuid);
         Content::query()->create([
-            'content_uuid'=>$location->uuid,
-            'user_uuid'=>$request->user_uuid,
+            'content_uuid' => $location->uuid,
+            'user_uuid' => $request->user_uuid,
         ]);
         if ($request->hasFile('images')) {
             foreach ($request->images as $item) {
                 UploadImage($item, Location::PATH_LOCATION, Location::class, $location->uuid, false, null, Upload::IMAGE);
             }
         }
+
         return response()->json([
             'item_added'
         ]);
@@ -53,63 +55,82 @@ class LocationController extends Controller
 
     public function update(Request $request)
     {
-        $location = Location::findOrFail($request->uuid);
+
+        $location = Location::query()->withoutGlobalScope('status')->findOrFail($request->uuid);
 
         $rules = [
             'name' => 'required|string|max:36',
             'price' => 'required|int',
             'details' => 'required',
-            'user_uuid'=>'required|exists:users,uuid',
+            'user_uuid' => 'required|exists:users,uuid',
             'category_contents_uuid' => 'required|exists:category_contents,uuid',
         ];
 
         $this->validate($request, $rules);
-        $location->update($request->only('name','details','price','user_uuid','category_contents_uuid'));
-        return response()->json([
-            'item_edited'
-        ]);
+        $location->update($request->only('name', 'details', 'price', 'user_uuid'));
+        $location->categories()->sync($request->category_contents_uuid);
 
-    }
+        if ($request->has('delete_images')) {
+
+            $images = Upload::query()->where('imageable_type',Location::class)->where('imageable_id',$location->uuid)->whereNotIn('uuid', $request->delete_images)->get();
+            foreach ($images as $item) {
+                File::delete(public_path(Location::PATH_LOCATION . $item->filename));
+                $item->delete();
+            }
+        }
+            if ($request->hasFile('images')) {
+                foreach ($request->images as $item) {
+                    UploadImage($item, Location::PATH_LOCATION, Location::class, $location->uuid, false, null, Upload::IMAGE);
+                }
+            }
+            return response()->json([
+                'item_edited'
+            ]);
+
+        }
 
     public function destroy($uuid)
     {
 
-        try {
-            $uuids=explode(',', $uuid);
-            $location=  Location::whereIn('uuid', $uuids)->get();
+//        try {
+        $uuids = explode(',', $uuid);
+        $location = Location::whereIn('uuid', $uuids)->withoutGlobalScope('status')->get();
 
-            foreach ($location as $item){
-                File::delete(public_path(Location::PATH_LOCATION.$item->imageLocation->filename));
-                $item->imageLocation()->delete();
-                $item->delete();
+        foreach ($location as $item) {
+            foreach ($item->imageLocation as $image) {
+                File::delete(public_path(Location::PATH_LOCATION . $image->filename));
+                $image->delete();
             }
-            return response()->json([
-                'item_deleted'
-            ]);
-        }catch (\Exception $e){
-            return response()->json([
-                'err'
-            ]);
+
+            $item->delete();
         }
+        return response()->json([
+            'item_deleted'
+        ]);
+//        }catch (\Exception $e){
+//            return response()->json([
+//                'err'
+//            ]);
+//        }
     }
 
     public function indexTable(Request $request)
     {
-        $location = Location::query()->withoutGlobalScope('status');
+        $location = Location::query()->withoutGlobalScope('status')->orderByDesc('created_at');
         return Datatables::of($location)
             ->filter(function ($query) use ($request) {
-                if ($request->status){
-                    $query->where('status',$request->status);
+                if ($request->status) {
+                    $query->where('status', $request->status);
                 }
-                if ($request->price){
-                    $query->where('price',$request->price);
+                if ($request->price) {
+                    $query->where('price', $request->price);
                 }
-                if ($request->name){
-                    $query->where('name',$request->name);
+                if ($request->name) {
+                    $query->where('name', $request->name);
                 }
 //
             })
-            ->addColumn('checkbox',function ($que){
+            ->addColumn('checkbox', function ($que) {
                 return $que->uuid;
             })
             ->addColumn('action', function ($que) {
@@ -119,9 +140,12 @@ class LocationController extends Controller
                 $data_attr .= 'data-price="' . $que->price . '" ';
                 $data_attr .= 'data-user_uuid="' . $que->user_uuid . '" ';
                 $data_attr .= 'data-details="' . $que->details . '" ';
-                $data_attr .= 'data-category_contents_uuid="' . $que->category_contents_uuid . '" ';
+                $data_attr .= 'data-images_uuid="' . implode(',', $que->imageLocation->pluck('uuid')->toArray()) .'" ';
+                $data_attr .= 'data-images="' . implode(',', $que->imageLocation->pluck('filename')->toArray()) .'" ';
 
-                $url = url('/locations/images/'.$que->uuid);
+                $data_attr .= 'data-category_contents_uuid="' . implode(',', $que->categories->pluck('uuid')->toArray()) . '," ';
+
+                $url = url('/locations/images/' . $que->uuid);
 
                 $string = '';
                 $string .= '<button class="edit_btn btn btn-sm btn-outline-primary btn_edit" data-toggle="modal"
@@ -129,20 +153,20 @@ class LocationController extends Controller
 
                 $string .= ' <button type="button" class="btn btn-sm btn-outline-danger btn_delete" data-uuid="' . $que->uuid .
                     '">' . __('delete') . '</button>';
-                $string .= ' <a href="'.$url.'"  class="btn btn-sm btn-outline-dark btn_image" data-uuid="' . $que->uuid .
+                $string .= ' <a href="' . $url . '"  class="btn btn-sm btn-outline-dark btn_image" data-uuid="' . $que->uuid .
                     '">' . __('images') . '  </a>';
-                $string .= ' <a href="'.$url.'"  class="btn btn-sm btn-outline-info btn_image" data-uuid="' . $que->uuid .
+                $string .= ' <a href="' . $url . '"  class="btn btn-sm btn-outline-info btn_image" data-uuid="' . $que->uuid .
                     '">' . __('details') . '  </a>';
                 return $string;
-            }) ->addColumn('status', function ($que)  {
+            })->addColumn('status', function ($que) {
                 $currentUrl = url('/');
-                if ($que->status==1){
-                    $data='
+                if ($que->status == 1) {
+                    $data = '
 <button type="button"  data-url="' . $currentUrl . "/locations/updateStatus/0/" . $que->uuid . '" id="btn_update" class=" btn btn-sm btn-outline-success " data-uuid="' . $que->uuid .
                         '">' . __('active') . '</button>
                     ';
-                }else{
-                    $data='
+                } else {
+                    $data = '
 <button type="button"  data-url="' . $currentUrl . "/locations/updateStatus/1/" . $que->uuid . '" id="btn_update" class=" btn btn-sm btn-outline-danger " data-uuid="' . $que->uuid .
                         '">' . __('inactive') . '</button>
                     ';
@@ -152,26 +176,30 @@ class LocationController extends Controller
             ->rawColumns(['action', 'status'])->toJson();
     }
 
-    public function updateStatus($status,$sup)
+    public function updateStatus($status, $sup)
     {
-        $uuids=explode(',', $sup);
+        $uuids = explode(',', $sup);
 
-        $location =  Location::query()->withoutGlobalScope('status')
-            ->whereIn('uuid',$uuids)
+        $location = Location::query()->withoutGlobalScope('status')
+            ->whereIn('uuid', $uuids)
             ->update([
-                'status'=>$status
+                'status' => $status
             ]);
         return response()->json([
             'item_edited'
         ]);
     }
-    public function imageIndex($uuid){
+
+
+    public function imageIndex($uuid)
+    {
         $uuid_location = $uuid;
         return view('admin.locations.images', compact('uuid_location'));
     }
-    public function imageIndexTable(Request $request,$uuid)
+
+    public function imageIndexTable(Request $request, $uuid)
     {
-        $images = Upload::query()->where('imageable_id',$uuid)->where('imageable_type',Location::class);
+        $images = Upload::query()->where('imageable_id', $uuid)->where('imageable_type', Location::class);
         return \Yajra\DataTables\DataTables::of($images)
             ->addColumn('checkbox', function ($que) {
                 return $que->uuid;
@@ -180,45 +208,50 @@ class LocationController extends Controller
                 $data_attr = '';
                 $data_attr .= 'data-uuid="' . $que->uuid . '" ';
                 $data_attr .= 'data-imageable_id="' . $que->imageable_id . '" ';
-                $data_attr .= 'data-image="' .url('/').Location::PATH_LOCATION.$que->filename . '" ';
+                $data_attr .= 'data-image="' . url('/') . Location::PATH_LOCATION . $que->filename . '" ';
                 $string = '';
-                $url = url('/dashboard/projects/images/'.$que->imageable_id);
+                $url = url('/dashboard/projects/images/' . $que->imageable_id);
                 $string .= '<button class="edit_btn btn btn-sm btn-outline-primary btn_edit" data-toggle="modal"
                     data-target="#edit_modal" ' . $data_attr . '>' . __('edit') . '</button>';
-                $string .= ' <button type="button" data-url="'.$url.'"  class="btn btn-sm btn-outline-danger btn_delete" data-uuid="' . $que->uuid .
-                    '">'.__('delete') .'  </button>';
+                $string .= ' <button type="button" data-url="' . $url . '"  class="btn btn-sm btn-outline-danger btn_delete" data-uuid="' . $que->uuid .
+                    '">' . __('delete') . '  </button>';
 
                 return $string;
-            })->addColumn('image',function ($que){
-                return url('/').Location::PATH_LOCATION.$que->filename;
+            })->addColumn('image', function ($que) {
+                return url('/') . Location::PATH_LOCATION . $que->filename;
             })
-
             ->rawColumns(['action'])->toJson();
     }
-    public function imageStore(Request $request){
-        foreach ($request->images as $item){
+
+    public function imageStore(Request $request)
+    {
+        foreach ($request->images as $item) {
             UploadImage($item, Location::PATH_LOCATION, Location::class, $request->uuid, false, null, Upload::IMAGE);
         }
         return response()->json("done");
     }
-    public function imageUpdate(Request $request){
+
+    public function imageUpdate(Request $request)
+    {
         UploadImage($request->images, Location::PATH_LOCATION, Location::class, $request->imageable_id, true, $request->uuid, Upload::IMAGE);
         return response()->json([
             'item_deleted'
         ]);
     }
-    public function imageDestroy($uuid,$uploade){
+
+    public function imageDestroy($uuid, $uploade)
+    {
 
         try {
-            $uuids=explode(',', $uploade);
-            $image=Upload::query()->whereIn('uuid',$uuids)->get();
-            foreach ($image as $item){
-                File::delete(public_path(Location::PATH_LOCATION.$item->filename));
+            $uuids = explode(',', $uploade);
+            $image = Upload::query()->whereIn('uuid', $uuids)->get();
+            foreach ($image as $item) {
+                File::delete(public_path(Location::PATH_LOCATION . $item->filename));
                 $item->delete();
             }
 
             return response()->json("done");
-        }catch (\Exception $e){
+        } catch (\Exception $e) {
             throw $e;
         }
     }
